Big update. Changes in order of importance.
- Wrote XML Parsing methods in User, so now it should be able to connect
and read messages at E-level! The only thing left to implement in User
for E-level is the method to write messages. After that, we need to
connect User and the GUI via the SessionController and write the Server,
but after that E-level functionality should work.
- Added OutgoingConnectionRequest as internal class to User
- Modified Message class so that it now takes more information
- Changed Notification method arguments in ConnectionObserver to String
and User (so it's like the old message).
- Reorganized Requests so now the User and Message is stored in the
superclass again.
- Added CryptoType enum and Crypto interface because I got so tired of
seeing the error messages in User